src/src/api/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
databases
psycopg2-binary
python-dotenv
pytest
pytest-asyncio
